# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.configField:
    properties:
      key:
        type: string
      label:
        type: string
      placeholder:
        type: string
      required:
        type: boolean
      secret:
        type: boolean
    type: object
  handler.configResponseDoc:
    properties:
      config:
        additionalProperties: true
        type: object
    type: object
  handler.configsResponse:
    properties:
      configs:
        items:
          additionalProperties: true
          type: object
        type: array
    type: object
...
  handler.errorResponse:
    properties:
      error:
        example: 请求失败
        type: string
    type: object
...
  handler.messageResponse:
    properties:
      message:
        example: 操作成功
        type: string
    type: object
  handler.picbedConfigRequest:
    properties:
      config:
        additionalProperties:
          type: string
        type: object
      config_name:
        type: string
      is_default:
        type: boolean
      picbed_type:
        type: string
    type: object
  handler.picbedTypeDef:
    properties:
      description:
        type: string
      fields:
        items:
          $ref: '#/definitions/handler.configField'
        type: array
      label:
        type: string
      value:
        type: string
    type: object
  handler.picbedTypesResponse:
    properties:
      types:
        items:
          $ref: '#/definitions/handler.picbedTypeDef'
        type: array
    type: object
...
info:
  contact: {}
  description: Markdown 图床批量转换平台 API
  title: PicBed Switcher API
  version: "1.0"
paths:
...
  /api/picbed/configs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.configsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取所有图床配置
      tags:
      - picbed
    post:
      consumes:
      - application/json
      parameters:
      - description: 图床配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.picbedConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.configResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 添加图床配置
      tags:
      - picbed
  /api/picbed/configs/{id}:
    delete:
      parameters:
      - description: 配置 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 删除图床配置
      tags:
      - picbed
    put:
      consumes:
      - application/json
      parameters:
      - description: 配置 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 图床配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.picbedConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.configResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新图床配置
      tags:
      - picbed
  /api/picbed/configs/{id}/default:
    put:
      parameters:
      - description: 配置 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.configResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 设置默认图床配置
      tags:
      - picbed
  /api/picbed/configs/{id}/test:
    post:
      parameters:
      - description: 配置 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 测试已保存的图床配置
      tags:
      - picbed
  /api/picbed/configs/test:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图床配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.picbedConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 测试未保存的图床配置
      tags:
      - picbed
  /api/picbed/types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.picbedTypesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: 获取支持的图床类型与配置字段
      tags:
      - picbed
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"